moved utils to pkg
